Extract NovelAI Settings
